pyproject.toml
[tool.poetry]
name = "dementia-detector"
version = "0.1.0"
description = ""
authors = ["tamil-phy <classicmecha@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.3"
langchain-openai = "^0.2.3"
langchain = "^0.3.4"
langchain-ollama = "^0.2.0"
python-dotenv = "^1.0.1"
uvicorn = "^0.32.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
